Fix story+routing
