+++ .storybook/main.ts
@@ -6,8 +6,7 @@ const config: StorybookConfig = {
         '@storybook/addon-links',
         '@storybook/addon-essentials',
         '@storybook/addon-onboarding',
-        '@storybook/addon-interactions',
-        'storybook-addon-react-router-v6'
+        '@storybook/addon-interactions'
     ],
     framework: {
         name: '@storybook/react-vite',
