+++ package.json
@@ -108,7 +108,6 @@
     "prettier-plugin-import-sort": "0.0.7",
     "prettier-plugin-tailwindcss": "^0.5.4",
     "storybook": "^7.4.5",
-    "storybook-addon-react-router-v6": "^2.0.7",
     "tailwindcss": "^3.3.3",
     "typescript": "^5.2.2",
     "vitest": "^0.31.1"
