+++ src/App.tsx
@@ -1,18 +1,11 @@
 import cx from 'classnames';
-import React from 'react';
-import { Navigate, RouterProvider, useParams } from 'react-router';
-import { createBrowserRouter, createHashRouter, useSearchParams } from 'react-router-dom';
+import React, { useEffect } from 'react';
 
-import { stripTrailingSlash } from './api/UseApiBasePath';
-import { USE_HASH_ROUTER, erInternFlate } from './constants';
+import { erInternFlate } from './constants';
 import { UppdateEventHandler } from './utils/UpdateEvent';
-import { Aktivitetskort } from './view/aktivitet/Aktivitetskort';
-import AppBody from './view/AppBody';
-import Dialog from './view/dialog/Dialog';
-import DialogInfoMelding from './view/dialog/DialogInfoMelding';
-import NyDialog from './view/dialog/NyDialog';
 import { Provider } from './view/Provider';
 import StatusAdvarsel from './view/statusAdvarsel/StatusAdvarsel';
+import { Routes } from './routes';
 
 interface Props {
     fnr?: string;
@@ -20,76 +13,9 @@ interface Props {
     visAktivitetDefault?: boolean;
 }
 
-const RedirectToDialogWithoutFnr = () => {
-    // /:fnr/:dialogId -> /:dialogId
-    const params = useParams();
-    return <Navigate replace to={`/` + params.dialogId} />;
-};
-const RedirectToNyDialogWithoutFnr = () => {
-    // Handle route /:fnr/ny?aktivitetId=<id> -> /ny?aktivitetId=<id>
-    const [queryParams, _] = useSearchParams();
-    const aktivitetId = queryParams.get('aktivitetId');
-    const queryPart = aktivitetId ? '?aktivitetId=' + aktivitetId : '';
-    return <Navigate replace to={`/ny${queryPart}`} />;
-};
-
-export const dialogRoutes = [
-    {
-        path: '/',
-        element: <AppBody />,
-        children: [
-            {
-                path: 'ny',
-                element: (
-                    <>
-                        <NyDialog />
-                        <Aktivitetskort />
-                    </>
-                )
-            },
-            {
-                path: ':dialogId',
-                element: (
-                    <>
-                        <Dialog />
-                        <Aktivitetskort />
-                    </>
-                )
-            },
-            {
-                path: '',
-                element: <DialogInfoMelding />
-            },
-            {
-                path: ':fnr/ny',
-                element: <RedirectToNyDialogWithoutFnr />
-            },
-            {
-                path: ':fnr/:dialogId',
-                element: <RedirectToDialogWithoutFnr />
-            },
-            {
-                path: '*',
-                element: <Navigate to={'/'} />
-            }
-        ]
-    }
-];
-
-const Routes = () => {
-    if (USE_HASH_ROUTER) {
-        const hashRouter = createHashRouter(dialogRoutes);
-        return <RouterProvider router={hashRouter} />;
-    }
-    let basename = stripTrailingSlash(import.meta.env.BASE_URL);
-    if (erInternFlate) basename = `/`;
-    const browserRouter = createBrowserRouter(dialogRoutes, { basename });
-    return <RouterProvider router={browserRouter} />;
-};
-
 const App = (props: Props) => {
-    console.log(props);
     const { fnr, visAktivitetDefault } = props;
+    useEffect(() => {}, []);
     return (
         <Provider visAktivitetDefault={visAktivitetDefault} fnr={fnr} erVeileder={!!fnr}>
             <StatusAdvarsel />
@@ -100,7 +26,7 @@ const App = (props: Props) => {
                     'max-h-[calc(100vh-80px)] min-h-[calc(100vh-80px)]': !erInternFlate
                 })}
             >
-                {/*<Routes />*/}
+                <Routes />
             </div>
         </Provider>
     );
